Replace ARGV with an empty array before each spec run.
Some of the specs were leaving ARGV in a non-pristine state (such as
those in runner_spec.rb), which was then implicitly used in
Thor::Base.start et al. It happened to be the case that the order in
which rspec was loading and running the specs made this not a problem,
but if the specs were run in a randomized order it could be a problem.

Closes #171.