Push conditionals into set_color method.
Calling #set_color required the client code to check the argument and
 #can_display_colors?. Pushing these checks into #set_color makes it much
easier to reuse this method.

This means that #set_color will now behave differently on TTY and non-TTY
terminals, so the #tty? method of $stdout has been stubbed in the
Thor::Shell::Color specs to always return true.