Add `merge` action to file colision menu
Currently, if want to reflect the changed contents, only have to grasp the
contents of diff and reflect it manually later or overwrite once to check
the difference. This is a little inconvenient.

If merge is selected, the specified merge tool will be executed. By doing
this, can reflect changes while checking.

Merge tool is obtained from environment variables and git config.
I doubt that it is appropriate to get a value from git config, but I am
thinking that there may be useful.