Lazily require `readline`
The current implementation of readline has issues under Windows, and
those issues prevent thor from being used. For example, in our bundler
Windows CI, we had to patch `rb-readline` to workaround some issues
because `rb-readline` (the default `readline` provider on Windows)
couldn't be required. Moreover, we still get some issues when requiring
`rb-readline` like

```
HOME environment variable (or HOMEDRIVE and HOMEPATH) must be set and point to a directory
```

Since in some situations the part of thor using `readline`
(`Thor::LineEditor`) is not required, for example, in most of CI
situations. I thought it could be a good idea to lazily require
`readline`.