Allow for nested `no_commands` blocks.
In certain circumstances we may wish to nest `no_commands` blocks.
For example, if we want to import any module which includes both methods
and `attr_reader` invocations.

However, since when exiting the no_commands block the @no_commands
instance var is reset to `false` we get unexpected behaviour.

This change uses a `NestedContext` object to track the depth of the
no_command blocks and ensure that we only start creating commands again
once we've left all of them.